Limit faction page to show single version
