+++ src/AppBundle/Controller/SearchController.php
@@ -370,7 +370,7 @@ class SearchController extends Controller
         // reconstruction de la bonne chaine de recherche pour affichage
         $q = $cardsData->buildQueryFromConditions($conditions);
         $rows = $cardsData->get_search_rows($conditions, $sort, $locale);
-        $rows = $this->select_only_latest_cards($rows);
+        $rows = $cardsData->select_only_latest_cards($rows);
         if ($q && $rows) {
             if (count($rows) == 1) {
                 $view = 'zoom';
@@ -500,21 +500,6 @@ class SearchController extends Controller
             "metadescription" => $meta,
             "locales"         => $locales,
         ], $response);
-    }
-
-    public function select_only_latest_cards(array $matchingCards)
-    {
-        $latestCardsByTitle = [];
-        foreach ($matchingCards as $card) {
-            $latestCard = null;
-            if (isset($latestCardsByTitle[$card->getTitle()])) {
-                $latestCard = $latestCardsByTitle[$card->getTitle()];
-            }
-            if (!$latestCard || $card->getCode() > $latestCard->getCode()) {
-                $latestCardsByTitle[$card->getTitle()] = $card;
-            }
-        }
-        return array_values($latestCardsByTitle);
     }
 
     /**
